More adjustments to current charts in separate components. #483.
