+++ src/views/reports/components/items-per-term-block.vue
@@ -145,10 +145,13 @@ export default {
             },
             immediate: true
         },
-        selectedTaxonomy() {
-            this.termsDisplayedPage = 1;
-            if (this.selectedTaxonomy && this.selectedTaxonomy.id)
-                this.loadTaxonomyTerms();
+        selectedTaxonomy: {
+            handler() {
+                this.termsDisplayedPage = 1;
+                if (this.selectedTaxonomy && this.selectedTaxonomy.id)
+                    this.loadTaxonomyTerms();
+            },
+            immediate: true
         },
         termsDisplayedPage() {
             this.buildTaxonomyTermsChart();
