+++ src/views/reports/js/reports-main.js
@@ -21,6 +21,7 @@ import TermsPerTaxonomyBlock from '../components/terms-per-taxonomy-block.vue';
 import MetadataTypesBlock from '../components/metadata-types-block.vue';
 import MetadataDistributionBlock from '../components/metadata-distribution-block.vue';
 import MetadataListBlock from '../components/metadata-list-block.vue';
+import CollectionsListBlock from '../components/collections-list-block.vue';
 
 Vue.use(VueApexCharts)
 
@@ -47,6 +48,7 @@ Vue.component('terms-per-taxonomy-block', TermsPerTaxonomyBlock);
 Vue.component('metadata-types-block', MetadataTypesBlock);
 Vue.component('metadata-distribution-block', MetadataDistributionBlock);
 Vue.component('metadata-list-block', MetadataListBlock);
+Vue.component('collections-list-block', CollectionsListBlock);
 Vue.component('apexchart', VueApexCharts);
 
 // Changing title of pages
